+++ tests/unit/type/NeverTypeTest.php
@@ -80,6 +80,7 @@ final class NeverTypeTest extends TestCase
         $this->assertFalse($type->isObject());
         $this->assertFalse($type->isSimple());
         $this->assertFalse($type->isStatic());
+        $this->assertFalse($type->isTrue());
         $this->assertFalse($type->isUnion());
         $this->assertFalse($type->isUnknown());
         $this->assertFalse($type->isVoid());