+++ tests/unit/type/UnknownTypeTest.php
@@ -72,6 +72,7 @@ final class UnknownTypeTest extends TestCase
         $this->assertFalse($this->type->isObject());
         $this->assertFalse($this->type->isSimple());
         $this->assertFalse($this->type->isStatic());
+        $this->assertFalse($this->type->isTrue());
         $this->assertFalse($this->type->isUnion());
         $this->assertTrue($this->type->isUnknown());
         $this->assertFalse($this->type->isVoid());